+++ lib/models/polygraph/variables/variable.dart
@@ -1,11 +1,41 @@
 library models.polygraph.variables.variable;
 
+import 'package:date/date.dart';
+import 'package:flutter_quiver/models/polygraph/transforms/transform.dart';
+import 'package:flutter_quiver/models/polygraph/variables/slope_intercept_variable.dart';
+import 'package:timeseries/timeseries.dart';
+import 'package:flutter/material.dart' hide Transform;
 
 mixin PolygraphVariable {
+  ///
   late final String name;
+
+  /// Internal representation used for the cache key
+  // String id();
+
+  /// What gets applied to this variable
+  final transforms = <Transform>[];
+
+  /// Data that gets displayed on the screen after all the transforms are
+  /// applied.  This method will access the cache except for trivial variables
+  /// like [SlopeInterceptVariable].
+  TimeSeries<num> timeSeries(Term term);
+
+  /// What gets displayed on the screen
   String label();
+  bool isMouseOver = false;
+  /// Should the variable be displayed on the plot?
+  bool isHidden = false;
+  Color? color;
+
+
+  /// How it's going to be persisted to the database
   Map<String,dynamic> toJson();
 }
+
+
+
+
 
 class TimeVariable extends Object with PolygraphVariable {
   TimeVariable({this.skipWeekends = false}) {
@@ -24,6 +54,12 @@ class TimeVariable extends Object with PolygraphVariable {
 
   @override
   String label() => 'Time';
+
+  @override
+  TimeSeries<num> timeSeries(Term term) {
+    // TODO: implement timeSeries
+    throw UnimplementedError();
+  }
 }