+++ lib/screens/polygraph/polygraph.dart
@@ -2,16 +2,12 @@ library screens.polygraph.polygraph;
 
 import 'package:date/date.dart';
 import 'package:flutter/material.dart' hide Interval;
-import 'package:flutter/widgets.dart' hide Interval;
-import 'package:flutter_quiver/models/common/experimental/power_deliverypoint_model.dart';
-import 'package:flutter_quiver/models/common/experimental/select_variable_model.dart';
-import 'package:flutter_quiver/models/common/market_model.dart';
-import 'package:flutter_quiver/models/common/region_model.dart';
-import 'package:flutter_quiver/models/common/term_model.dart';
+import 'package:flutter_quiver/main.dart';
 import 'package:flutter_quiver/models/polygraph/polygraph_model.dart';
-import 'package:flutter_quiver/screens/polygraph/polygraph_ui.dart';
+import 'package:flutter_quiver/models/polygraph/variables/variable_display_config.dart';
+import 'package:flutter_quiver/models/polygraph/variables/variable_selection.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter_web_plotly/flutter_web_plotly.dart';
 import 'package:timezone/timezone.dart';
 
 class Polygraph extends ConsumerStatefulWidget {
@@ -28,8 +24,12 @@ class _PolygraphState extends ConsumerState<Polygraph> {
   final focusNodeTerm = FocusNode();
   late ScrollController _scrollControllerV;
   late ScrollController _scrollControllerH;
+  late Plotly plotly;
+
 
   String? _errorTerm;
+
+  final variableSelection = VariableSelection();
 
   @override
   void initState() {
@@ -53,6 +53,13 @@ class _PolygraphState extends ConsumerState<Polygraph> {
         });
       }
     });
+
+    var aux = DateTime.now().hashCode;
+    plotly = Plotly(
+      viewId: 'polygraph-div-$aux',
+      data: const [],
+      layout: PolygraphState.layout,
+    );
   }
 
   @override
@@ -64,9 +71,11 @@ class _PolygraphState extends ConsumerState<Polygraph> {
     super.dispose();
   }
 
-
   @override
   Widget build(BuildContext context) {
+    var state = ref.watch(providerOfPolygraph);
+    var categories = variableSelection
+        .getCategoriesForLevel(variableSelection.categories.length);
 
     return Scaffold(
       appBar: AppBar(
@@ -83,10 +92,35 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                           width: 500,
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              TextButton(
+                                  onPressed: () {},
+                                  child: const Text('Load/Save')),
+                            ],
+                          ),
+                        )
+                      ],
+                      contentPadding: const EdgeInsets.all(12),
+                    );
+                  });
+            },
+            icon: const Icon(Icons.archive_outlined),
+            tooltip: 'Load/Save page',
+          ),
+          IconButton(
+            onPressed: () {
+              showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return SimpleDialog(
+                      children: [
+                        SizedBox(
+                          width: 500,
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
                             children: const [
-                              Text(
-                                  'Experimental UI for curve visualization.'
-                                      '\n'),
+                              Text('Experimental UI for curve visualization.'
+                                  '\n'),
                             ],
                           ),
                         )
@@ -97,7 +131,7 @@ class _PolygraphState extends ConsumerState<Polygraph> {
             },
             icon: const Icon(Icons.info_outline),
             tooltip: 'Info',
-          )
+          ),
         ],
       ),
       body: Padding(
@@ -114,7 +148,8 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                 const SizedBox(
                   height: 8,
                 ),
-                /// Historical term
+
+                /// Term
                 SizedBox(
                     width: 140,
                     child: TextFormField(
@@ -122,11 +157,11 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                       decoration: InputDecoration(
                         labelText: 'Term',
                         labelStyle:
-                        TextStyle(color: Theme.of(context).primaryColor),
+                            TextStyle(color: Theme.of(context).primaryColor),
                         helperText: '',
                         enabledBorder: UnderlineInputBorder(
                           borderSide:
-                          BorderSide(color: Theme.of(context).primaryColor),
+                              BorderSide(color: Theme.of(context).primaryColor),
                         ),
                         errorText: _errorTerm,
                       ),
@@ -149,6 +184,79 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                 const SizedBox(
                   height: 8,
                 ),
+
+                if (variableSelection.categories.isNotEmpty)
+                  Wrap(
+                    spacing: 5.0,
+                    crossAxisAlignment: WrapCrossAlignment.center,
+                    children: [
+                      const Text('Selection '),
+                      ...List.generate(variableSelection.categories.length,
+                          (index) {
+                        return InputChip(
+                          label: Text(
+                            variableSelection.categories[index],
+                          ),
+                          onDeleted: () {
+                            setState(() {
+                              variableSelection.removeFromLevel(index);
+                            });
+                          },
+                          deleteIcon: const Icon(Icons.close),
+                          backgroundColor: MyApp.background,
+                        );
+                      })
+                    ],
+                  ),
+                const SizedBox(
+                  height: 16,
+                ),
+                if (!variableSelection.isSelectionDone())
+                  const Text('Choose a category'),
+                const SizedBox(
+                  height: 8,
+                ),
+                if (!variableSelection.isSelectionDone())
+                  Wrap(
+                    spacing: 5.0,
+                    children: List<Widget>.generate(
+                      categories.length,
+                      (int index) {
+                        return ChoiceChip(
+                          selectedColor: MyApp.background,
+                          label: Text(categories[index]),
+                          selected: false,
+                          onSelected: (bool selected) {
+                            setState(() {
+                              if (selected) {
+                                variableSelection
+                                    .addCategory(categories[index]);
+                              }
+                            });
+                          },
+                        );
+                      },
+                    ).toList(),
+                  ),
+                const SizedBox(
+                  height: 48,
+                ),
+                if (variableSelection.isSelectionDone())
+                  ElevatedButton(
+                    onPressed: () {
+                      setState(() {
+                        variableSelection.removeFromLevel(0);
+                      });
+                    },
+                    child: const Text('OK'),
+                    // style: ElevatedButton.styleFrom(
+                    //     backgroundColor: Colors.blue[700],
+                    //     foregroundColor: Colors.white),
+                  ),
+
+                ///
+                /// The variables
+                ///
                 Wrap(
                   direction: Axis.horizontal,
                   spacing: 36,
@@ -162,19 +270,19 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                           child: Text(
                             'X axis',
                             style:
-                            TextStyle(color: Colors.blueGrey, fontSize: 16),
+                                TextStyle(color: Colors.blueGrey, fontSize: 16),
                           ),
                         ),
                         TextButton(
                           onPressed: () {},
                           child: Text(
-                            variableModel.xAxisLabel(),
+                            state.xVariable.name,
                             style: const TextStyle(
                                 color: Colors.black, fontSize: 16),
                           ),
                           style: TextButton.styleFrom(
-                            padding:
-                            const EdgeInsets.only(left: 0, top: 0, bottom: 0),
+                            padding: const EdgeInsets.only(
+                                left: 0, top: 0, bottom: 0),
                             alignment: Alignment.centerLeft,
                             // backgroundColor: Colors.orange,
                           ),
@@ -187,7 +295,8 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                       children: [
                         const Text(
                           'Y axis',
-                          style: TextStyle(color: Colors.blueGrey, fontSize: 16),
+                          style:
+                              TextStyle(color: Colors.blueGrey, fontSize: 16),
                         ),
                         //
                         //
@@ -195,18 +304,16 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                         //
                         //
                         ...[
-                          for (var i = 0; i < ys.length; i++)
+                          for (var i = 0; i < state.yVariables.length; i++)
                             MouseRegion(
                               onEnter: (_) {
                                 setState(() {
-                                  variableModel.yVariablesHighlightStatus[i] =
-                                  true;
+                                  state.yVariables[i].isMouseOver = true;
                                 });
                               },
                               onExit: (_) {
                                 setState(() {
-                                  variableModel.yVariablesHighlightStatus[i] =
-                                  false;
+                                  state.yVariables[i].isMouseOver = false;
                                 });
                               },
                               child: Stack(
@@ -215,73 +322,65 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                                   TextButton(
                                     onPressed: () {},
                                     child: Text(
-                                      variableModel.yAxisLabel(i),
-                                      style: const TextStyle(
-                                          color: Colors.black, fontSize: 16),
+                                      state.yVariables[i].label(),
+                                      style: TextStyle(
+                                          color: state.yVariables[i].color ?? VariableDisplayConfig.defaultColors[i],
+                                          decoration: state.yVariables[i].isHidden ? TextDecoration.lineThrough : TextDecoration.none,
+                                          fontSize: 16),
                                     ),
                                     style: TextButton.styleFrom(
                                       padding: const EdgeInsets.only(
-                                          left: 0, right: 90),
+                                          left: 0, right: 120),
                                       // minimumSize: Size(200, 24),
                                     ),
                                   ),
 
                                   /// show the icons only on hover ...
-                                  if (variableModel.yVariablesHighlightStatus[i])
+                                  if (state.yVariables[i].isMouseOver)
                                     Row(
                                       children: [
                                         IconButton(
                                           tooltip: 'Edit',
                                           onPressed: () async {
                                             setState(() {
-                                              variableModel.editedIndex = i + 1;
+                                              // variableModel.editedIndex = i + 1;
                                             });
                                             await showDialog(
                                                 barrierDismissible: false,
                                                 context: context,
                                                 builder: (context) {
-                                                  return ChangeNotifierProvider
-                                                      .value(
-                                                      value: variableModel,
-                                                      builder:
-                                                          (context, _) =>
-                                                          AlertDialog(
-                                                            elevation:
-                                                            24.0,
-                                                            title: const Text(
-                                                                'Select'),
-                                                            content:
-                                                            Column(
-                                                              mainAxisSize:
-                                                              MainAxisSize
-                                                                  .min,
-                                                              children: const [
-                                                                EditorMain(),
-                                                              ],
-                                                            ),
-                                                            actions: [
-                                                              TextButton(
-                                                                  child: const Text(
-                                                                      'CANCEL'),
-                                                                  onPressed:
-                                                                      () {
-                                                                    /// ignore changes the changes
-                                                                    Navigator.of(context)
-                                                                        .pop();
-                                                                  }),
-                                                              ElevatedButton(
-                                                                  child: const Text(
-                                                                      'OK'),
-                                                                  onPressed:
-                                                                      () {
-                                                                    /// harvest the values
-                                                                    Navigator.of(context)
-                                                                        .pop();
-                                                                  }),
-                                                            ],
-                                                          ));
+                                                  return AlertDialog(
+                                                      elevation: 24.0,
+                                                      title:
+                                                          const Text('Select'),
+                                                      content: Column(
+                                                        mainAxisSize:
+                                                            MainAxisSize.min,
+                                                        children: const [
+                                                          Text('Boo'),
+                                                        ],
+                                                      ),
+                                                      actions: [
+                                                        TextButton(
+                                                            child: const Text(
+                                                                'CANCEL'),
+                                                            onPressed: () {
+                                                              /// ignore changes the changes
+                                                              Navigator.of(
+                                                                      context)
+                                                                  .pop();
+                                                            }),
+                                                        ElevatedButton(
+                                                            child: const Text(
+                                                                'OK'),
+                                                            onPressed: () {
+                                                              /// harvest the values
+                                                              Navigator.of(
+                                                                      context)
+                                                                  .pop();
+                                                            }),
+                                                      ]);
                                                 });
-                                            print(variableModel.yAxisVariables());
                                           },
                                           visualDensity: VisualDensity.compact,
                                           constraints: const BoxConstraints(),
@@ -297,7 +396,7 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                                           tooltip: 'Remove',
                                           onPressed: () {
                                             setState(() {
-                                              variableModel.removeVariableAt(i);
+                                              // variableModel.removeVariableAt(i);
                                             });
                                           }, // delete the sucker
                                           visualDensity: VisualDensity.compact,
@@ -314,7 +413,7 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                                           tooltip: 'Copy',
                                           onPressed: () {
                                             setState(() {
-                                              variableModel.copy(i);
+                                              // variableModel.copy(i);
                                             });
                                           }, // delete the sucker
                                           visualDensity: VisualDensity.compact,
@@ -323,6 +422,23 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                                               left: 0, right: 8),
                                           icon: Icon(
                                             Icons.copy,
+                                            color: Colors.blueGrey[300],
+                                            size: 20,
+                                          ),
+                                        ),
+                                        IconButton(
+                                          tooltip: 'Hide/Show',
+                                          onPressed: () {
+                                            setState(() {
+                                              state.yVariables[i].isHidden = !state.yVariables[i].isHidden;
+                                            });
+                                          },
+                                          visualDensity: VisualDensity.compact,
+                                          constraints: const BoxConstraints(),
+                                          padding: const EdgeInsets.only(
+                                              left: 0, right: 8),
+                                          icon: Icon(
+                                            Icons.format_strikethrough,
                                             color: Colors.blueGrey[300],
                                             size: 20,
                                           ),
@@ -337,86 +453,13 @@ class _PolygraphState extends ConsumerState<Polygraph> {
                     ),
                   ],
                 ),
-
                 const SizedBox(
                   height: 12,
                 ),
+                /// The chart
 
 
 
-                // FutureBuilder(
-                //   future:
-                //   RateBoardState.getOffers(state.region, state.stateName),
-                //   builder: (context, snapshot) {
-                //     List<Widget> children;
-                //     if (snapshot.hasData) {
-                //       var columns = _makeColumns(state);
-                //       var tbl = state.makeOfferTable(
-                //           asOfDate: Date.today(location: UTC).previous);
-                //       // print(tbl.take(2).map((e) => e.toMap()));
-                //       if (tbl.isEmpty) {
-                //         children = [const Text('')];
-                //       } else {
-                //         children = [
-                //           SizedBox(
-                //             width: 1000,
-                //             child: PaginatedDataTable(
-                //               dataRowHeight: 64,
-                //               columnSpacing: 24,
-                //               columns: columns,
-                //               source: _DataTableSource(tbl),
-                //               rowsPerPage: min(20, tbl.length),
-                //               showFirstLastButtons: true,
-                //               header: const Text('Current offers'),
-                //               actions: [
-                //                 IconButton(
-                //                     onPressed: () {
-                //                       Clipboard.setData(ClipboardData(
-                //                           text: table.Table.from(
-                //                               tbl.map((e) => e.toMap()))
-                //                               .toCsv()));
-                //                     },
-                //                     tooltip: 'Copy',
-                //                     icon: const Icon(Icons.content_copy)),
-                //                 IconButton(
-                //                     onPressed: () => downloadTableToCsv(
-                //                         tbl.map((e) => e.toMap()).toList()),
-                //                     tooltip: 'Download',
-                //                     icon: const Icon(Icons.download_outlined))
-                //               ],
-                //               // )
-                //             ),
-                //           ),
-                //           if (state.stateName == 'MA')
-                //             const Text('*A mention of 100% in plan Features indicates '
-                //                 'that the plan is supplied with 100% green power. '
-                //               , style: TextStyle(fontStyle: FontStyle.italic),),
-                //         ];
-                //       }
-                //     } else if (snapshot.hasError) {
-                //       children = [
-                //         const Icon(Icons.error_outline, color: Colors.red),
-                //         Text(
-                //           snapshot.error.toString(),
-                //           style: const TextStyle(fontSize: 16),
-                //         )
-                //       ];
-                //     } else {
-                //       children = [
-                //         const SizedBox(
-                //             height: 40,
-                //             width: 40,
-                //             child: CircularProgressIndicator(
-                //               strokeWidth: 2,
-                //             ))
-                //       ];
-                //     }
-                //     return Column(
-                //         crossAxisAlignment: CrossAxisAlignment.start,
-                //         // mainAxisSize: MainAxisSize.min,
-                //         children: children);
-                //   },
-                // ),
               ],
             ),
           ),
@@ -424,7 +467,5 @@ class _PolygraphState extends ConsumerState<Polygraph> {
         ),
       ),
     );
-
-
   }
 }